9.1.3 add internet available check into the provider. If internet is unavailable, it will reject the promises

+++ src/api/provider.js
@@ -1,21 +1,45 @@
+const isOnLine = () => {
+  return window.navigator.onLine;
+};
+
 export default class Provider {
   constructor(api) {
     this._api = api;
   }
 
   getTasks() {
-    return this._api.getTasks();
+    if (isOnLine()) {
+      return this._api.getTasks();
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   createTask(task) {
-    this._api.createTask(task);
+    if (isOnLine()) {
+      return this._api.createTask(task);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   updateTask(id, task) {
-    this._api.updateTask(id, task);
+    if (isOnLine()) {
+      return this._api.updateTask(id, task);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   deleteTask(id) {
-    this._api.deleteTask(id);
+    if (isOnLine()) {
+      return this._api.deleteTask(id);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 }
